Break apart explore page
