+++ pages/explore/utils.ts
@@ -1,0 +1,6 @@
+export function deletePins(str) {
+  let previous = document.querySelectorAll(str);
+  previous.forEach((marker) => {
+    marker.remove();
+  });
+}
